+++ src/LCEMU/gui/gui_utils.h
@@ -5,5 +5,5 @@
 HWND UtilCreateWindow(const TCHAR* win_class, const TCHAR* title, const WNDPROC& proc, const COLORREF& Background, const UINT style);
 void MessageLoop(const HWND& hwnd);
 
-bool FixWindowCornor(const HWND& hwnd);
+bool FixWindowCorner(const HWND& hwnd);
 std::pair<int, int> CalcWindowSize(const HWND& hwnd, const int& client_w, const int& client_h);